requirements.txt
albumentations==1.1.0
altair==4.1.0
anyio==3.4.0
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==21.12b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
entrypoints==0.3
fastapi==0.70.1
gitdb==4.0.9
GitPython==3.1.24
h11==0.12.0
idna==3.3
imageio==2.13.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.0
jupyterlab==3.2.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.21.4
olefile==0.46
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
pyarrow==6.0.1
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.10.0
Pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.26.0
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.16.0
streamlit==1.2.0
streamlit-cropper==0.1.3
streamlit-drawable-canvas==0.8.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.16.0
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.2
wheel==0.37.0
widgetsnbextension==3.5.2
wincertstore==0.2
zipp==3.6.0
